Moved routes dir into /server to handle storing of image for image recognition
